display root element on minimap
